pyproject.toml
[project]
name = "neuroflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.24",
    "langchain-openai>=0.3.14",
    "openai-whisper>=20240930",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "streamlit-calendar>=1.2.1",
]
